Getting rid of the outer folder
