--- Automatic-Plant-Care-backup-age-v2/README.md
@@ -1,1 +1,0 @@
-# Automatic Plant Care
