--- Automatic-Plant-Care-backup-age-v2/web-server/public/index.html
@@ -1,89 +1,0 @@
-<!DOCTYPE html>
-<html lang='en'>
-<head>
-	<title>Automatic Plant Care</title>
-	<meta charset="UTF-8">
-	<link rel='stylesheet' href='./stylesheets/style.css'></link>
-</head>
-
-<body>
-<form>
-	<h1>Automatic Plant Care</h1>
-
-	<div id='content'>	
-		<div id='status-box'>
-			<h2>Plant's Status</h2>
-			<p>Moisture level: 
-				<code><strong><span id="moistStatus"></span></strong></code>
-			</p>
-			<p>Light level: 
-				<code><strong><span id="lightStatus"></span></strong></code>
-			</p>
-		</div>
-		
-		<div id='left-box'>
-			<div id='modediv'>
-				<h2>Plant Care Settings</h2>
-				<p>
-					Email: 
-					<input type="text"   id="userEmail" value="sample_email@gmail.com" size="17"/>
-				</p>
-				<p>
-					<input type="button" id="confirmEmail" value="Confirm"/>
-					<input type="button" id="sampleEmail" value="Send Sample Email"/>
-				</p>
-				<h3>Pump Control</h3>
-				<p>
-					Water pump status: <strong><span id="pumpStatus"></span></strong></br>
-				</p>
-				<p>
-					<input type="button" id="pumpOn"  value="Pump on"/>
-					<input type="button" id="pumpOff" value="Pump off"/>
-				</p>
-				<p>
-					<div id="range-sliders" class="range-sliders">
-						<h3>Moisture Safety Threshold</h3>
-						<p>Min:</p>
-						<div class="range-slider-group range-slider-group-moist-min">
-							<input type="range" min="0" max="100" value="50" class="range-slider range-slider-moist-min" id="range-slider-moist-min"/>
-							<input type="text" min="0" max="100" value="50" class="input-slider input-slider-moist-min" id="input-slider-moist-min" size="1"/ readonly>
-							% VMC
-						</div>
-						<p>Max:</p>
-						<div class="range-slider-group range-slider-group-moist-max">
-							<input type="range" min="0" max="100" value="50" class="range-slider range-slider-moist-max" id="range-slider-moist-max"/>
-							<input type="text" min="0" max="100" value="50" class="input-slider input-slider-moist-max" id="input-slider-moist-max" size="1"/ readonly>
-							% VMC
-						</div>
-
-						<h3>Light Safety Threshold</h3>
-						<p>Min:</p>
-						<div class="range-slider-group range-slider-group-light-min">
-							<input type="range" min="0" max="13000" value="1000" class="range-slider range-slider-light-min" id="range-slider-light-min"/>
-							<input type="text" min="0" max="13000" value="1000" class="input-slider input-slider-light-min" id="input-slider-light-min" size="1"/ readonly>
-							lux
-						</div>
-						<p>Max:</p>
-						<div class="range-slider-group range-slider-group-light-max">
-							<input type="range" min="0" max="13000" value="1000" class="range-slider range-slider-light-max" id="range-slider-light-max"/>
-							<input type="text" min="0" max="13000" value="1000" class="input-slider input-slider-light-max" id="input-slider-light-max" size="1"/ readonly>
-							lux
-						</div>
-					</div>
-				</p>
-		</div>
-		<div id='error-box'>
-			<h2>Server Error</h2>
-			<p>Server error detected. Please resolve error before continuing.</p>
-			<div><code id='error-text'>No error yet... did you hide this div correctly via Javascript? :)</code></div>
-		</div>
-
-	</div>
-</form>
-	
-	<script src='/socket.io/socket.io.js' type='text/javascript'></script>
-	<script src='http://code.jquery.com/jquery-1.11.1.min.js' type='text/javascript'></script>
-
-	<script src='javascripts/udp_ui.js' type='text/javascript'></script>
-</body>
-</html>
