--- Automatic-Plant-Care-backup-age-v2/web-server/public/stylesheets/style.css
@@ -1,64 +1,0 @@
-body {
-	width: 800px;
-	margin-left: auto;
-	margin-right: auto;
-}
-h1, h2 {
-	margin-bottom: 1px;
-	padding-bottom: 1px;
-}
-
-#content {
-	margin-top: 10pt;
-}
-
-#moistStatus {
-	font-size: 16px;
-}
-
-#lightStatus {
-	font-size: 16px;
-}
-
-#plantAge {
-  font-size: 16px;
-}
-
-#DOB {
-  font-size: 16px;
-}
-
-#status-box {
-	float: right;
-	width: 250px;
-	min-height: 200pt;
-	border: thin black solid;
-	background-color: #f3f3f3;
-	padding-left: 10pt;
-	padding-right: 10pt;
-}
-
-#left-box {
-	margin-right: 300px;	
-	border: thin black solid;
-	padding-left: 10pt;
-	padding-right: 10pt;
-	min-height: 200pt;
-}
-
-#error-box {
-	border: thin black solid;
-	clear: both;
-	background-color: red;
-	padding-left: 10pt;
-	padding-right: 10pt;
-	min-height: 100pt;
-	margin-top: 20px;
-	display: none;
-}
-
-.range-sliders .range-slider-group {
-	display: flex;
-	flex-wrap: wrap;
-	align-items: center;
-}
