--- Automatic-Plant-Care-backup-age-v2/web-server/server.js
@@ -1,70 +1,0 @@
-// After creating package.json, install modules:
-//   $ npm install
-//
-// Launch server with:
-//   $ node server.js
-var PORT_NUMBER = 8088;
-
-
-var http = require('http');
-var fs   = require('fs');
-var path = require('path');
-var mime = require('mime');
-
-/* 
- * Create the static web server
- */
-var server = http.createServer(function(request, response) {
-	var filePath = false;
-	
-	if (request.url == '/') {
-		filePath = 'public/index.html';
-	} else {
-		filePath = 'public' + request.url;
-	}
-	
-	var absPath = './' + filePath;
-	serveStatic(response, absPath);
-});
-
-server.listen(PORT_NUMBER, function() {
-	console.log("Server listeneing on port " + PORT_NUMBER);
-});
-
-function serveStatic(response, absPath) {
-	fs.exists(absPath, function(exists) {
-		if (exists) {
-			fs.readFile(absPath, function(err, data) {
-				if (err) {
-					send404(response);
-				} else {
-					sendFile(response, absPath, data);
-				}
-			});
-		} else {
-			send404(response);
-		}
-	});
-}
-
-function send404(response) {
-	response.writeHead(404, {'Content-Type': 'text/plain'});
-	response.write('Error 404: resource not found.');
-	response.end();
-}
-
-function sendFile(response, filePath, fileContents) {
-	response.writeHead(
-			200,
-			{"content-type": mime.lookup(path.basename(filePath))}
-		);
-	response.end(fileContents);
-}
-
-
-/*
- * Create the Userver to listen for the websocket
- */
-var udpServer = require('./lib/udp_server');
-udpServer.listen(server);
-
