+++ include/utils.h
@@ -9,5 +9,7 @@
 // Utility functions to read from or write to file
 int Utils_readFromFileToScreen(const char *fileName);
 void Utils_writeToFile(const char *fileName, const char *command);
+int Utils_read_bin(const char* fileName, char* buffer);
+int Utils_write_bin(const char* fileName, const char* value, int size);
 
 #endif
