+++ web-server/public/stylesheets/style.css
@@ -20,6 +20,14 @@ h1, h2 {
 	font-size: 16px;
 }
 
+#plantAge {
+  font-size: 16px;
+}
+
+#DOB {
+  font-size: 16px;
+}
+
 #status-box {
 	float: right;
 	width: 250px;
